requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
tensorflow==2.15.0
pillow==10.1.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
